refactor: Rework inplace upgrade as refresh
