+++ tox.ini
@@ -33,7 +33,7 @@ set_env =
     tls: TESTFILE=test_tls.py
     auth: TESTFILE=test_auth.py
     provider: TESTFILE=test_provider.py
-    inplace: TESTFILE=inplace_upgrades/test_inplace_upgrades.py
+    refresh: TESTFILE=refresh/test_refresh.py
 pass_env =
     PYTHONPATH
     CHARM_BUILD_DIR
@@ -95,7 +95,7 @@ commands =
 
 
 
-[testenv:integration-{charm,trust,ha,upgrade,external-access,tls,dynamic-allocation,observability,iceberg,metastore,auth,provider,inplace}]
+[testenv:integration-{charm,trust,ha,upgrade,external-access,tls,dynamic-allocation,observability,iceberg,metastore,auth,provider,refresh}]
 description = Run integration tests
 set_env =
     {[testenv]set_env}
